Sidebar working better
